+++ src/assets/events/halloween/assets/nfts/farmer_bath.png
