+++ src/assets/events/halloween/assets/nfts/homeless_tent.png
