+++ src/assets/events/halloween/assets/nfts/nyon_statue.png
