+++ src/assets/events/halloween/assets/nfts/potato_statue.png
