+++ src/assets/events/halloween/assets/nfts/rooster.gif
