+++ src/assets/events/halloween/assets/nfts/undead_chicken.gif
