+++ src/assets/events/halloween/assets/nfts/wicker_man_fire.png
