+++ src/features/halloween/game/components/Decorations.tsx
@@ -4,30 +4,30 @@
 import React from "react";
 
 //sunflower rock path edited to match event theme
-import sunflowerRock from "assets/nfts/halloween/sunflower_rock.png";
+import sunflowerRock from "assets/events/halloween/assets/nfts/sunflower_rock.png";
 import sunflowerTombstone from "assets/nfts/sunflower_tombstone.png";
 import sunflowerStatue from "assets/nfts/sunflower_statue.png";
 //potato statue path edited to match event theme
-import potatoStatue from "assets/nfts/halloween/potato_statue.png";
+import potatoStatue from "assets/events/halloween/assets/nfts/potato_statue.png";
 import christmasTree from "assets/nfts/christmas_tree.png";
 import dog from "assets/nfts/farm_dog.gif";
 import cat from "assets/nfts/farm_cat.gif";
 import gnome from "assets/nfts/gnome.gif";
 import goblinKing from "assets/nfts/goblin_king.png";
 //fountain path edited to match event theme
-import fountain from "assets/nfts/halloween/fountain.gif";
+import fountain from "assets/events/halloween/assets/nfts/fountain.gif";
 import goldenBonsai from "assets/nfts/golden_bonsai.png";
 //rooster path edited to match event theme
-import rooster from "assets/nfts/halloween/rooster.gif";
+import rooster from "assets/events/halloween/assets/nfts/rooster.gif";
 //undead chicken path edited to match event theme
-import undeadChicken from "assets/nfts/halloween/undead_chicken.gif";
+import undeadChicken from "assets/events/halloween/assets/nfts/undead_chicken.gif";
 import pottedSunflower from "assets/decorations/potted_sunflower.png";
 // mysterious head path edited to match event theme
-import mysteriousHead from "assets/nfts/halloween/mysterious_head.png";
+import mysteriousHead from "assets/events/halloween/assets/nfts/mysterious_head.png";
 // homeless tent path edited to match event theme
-import homelessTent from "assets/nfts/halloween/homeless_tent.png";
+import homelessTent from "assets/events/halloween/assets/nfts/homeless_tent.png";
 // farmerBath path edited to match event theme
-import farmerBath from "assets/nfts/halloween/farmer_bath.png";
+import farmerBath from "assets/events/halloween/assets/nfts/farmer_bath.png";
 import swimmer from "assets/npcs/swimmer.gif";
 import skullHand from "assets/decorations/skull_hand.png";
 import easterBunny from "assets/nfts/easter/easter_bunny_eggs.gif";
